* Completely redid everything * Moved to Module Layout for chat client * Moved FaceitMatch.FaceitTeam to its own class file * Lots and Lots of changes to things. * Version 1.0.4 release

+++ src/FaceitSharp.Core/IFaceitConfig.cs
@@ -21,11 +21,11 @@ public interface IFaceitSocketConfig
 {
     string Uri { get; }
 
-    int ReconnectTimeout { get; }
+    double ReconnectTimeout { get; }
 
-    int ReconnectTimeoutError { get; }
+    double ReconnectTimeoutError { get; }
 
-    int KeepAliveInterval { get; }
+    double KeepAliveInterval { get; }
 
     double ResponseTimeout { get; }
